Slight grammatical fixes on the README.
I've just bought my HORI Truck Control System to use on my Linux, so I will be invested in contributing to this repository. I'm doing this just to get acquainted and I hope my small grammatical fixes are not pedantic or intrusive. I also plan to add a PKGBUILD so that it can be easily built on Archlinux and added to the AUR.